refactor: split pkg/config/ast.go
