+++ pkg/config/ast.go
@@ -1,8 +1,6 @@
 package config
 
 import (
-	"encoding/json"
-	"encoding/xml"
 	"fmt"
 	"strings"
 )
@@ -141,93 +139,6 @@ func cloneNodes(nodes []*Node) []*Node {
 	return result
 }
 
-// ExpandGroups resolves all "apply-groups" references in the tree.
-// It collects group definitions from the "groups" stanza, then for each
-// "apply-groups <name>" node, clones the referenced group's children and
-// merges them into the parent. After expansion, both "groups" and
-// "apply-groups" nodes are removed from the tree.
-func (t *ConfigTree) ExpandGroups() error {
-	return t.expandGroups(false, nil)
-}
-
-// ExpandGroupsTagged is like ExpandGroups but tags each inherited node
-// with InheritedFrom set to the group name, for "| display inheritance".
-func (t *ConfigTree) ExpandGroupsTagged() error {
-	return t.expandGroups(true, nil)
-}
-
-// ExpandGroupsWithVars is like ExpandGroups but resolves ${var} references
-// in apply-groups names before lookup. This supports Junos-style per-node
-// group selection, e.g. apply-groups "${node}" with vars {"node": "node0"}.
-func (t *ConfigTree) ExpandGroupsWithVars(vars map[string]string) error {
-	return t.expandGroups(false, vars)
-}
-
-// resolveVars replaces ${key} placeholders in s with values from vars.
-func resolveVars(s string, vars map[string]string) string {
-	if vars == nil {
-		return s
-	}
-	for k, v := range vars {
-		s = strings.ReplaceAll(s, "${"+k+"}", v)
-	}
-	return s
-}
-
-func (t *ConfigTree) expandGroups(tagInherited bool, vars map[string]string) error {
-	// Collect group definitions: groups { <name> { ... } }
-	groups := make(map[string]*Node)
-	for _, child := range t.Children {
-		if child.Name() == "groups" {
-			for _, g := range child.Children {
-				if len(g.Keys) < 1 {
-					continue
-				}
-				name := g.Keys[0]
-				if len(g.Keys) > 1 {
-					name = g.Keys[1]
-				}
-				groups[name] = g
-			}
-		}
-	}
-
-	// If no groups defined, just strip any stale apply-groups references.
-	if len(groups) == 0 {
-		return t.stripApplyGroups(vars)
-	}
-
-	// Recursively resolve apply-groups at all levels.
-	// The nil ancestorPath means we're at the top level.
-	if err := expandGroupsRecursive(&t.Children, groups, nil, nil, tagInherited, vars); err != nil {
-		return err
-	}
-
-	// Remove the "groups" stanza itself.
-	filtered := make([]*Node, 0, len(t.Children))
-	for _, child := range t.Children {
-		if child.Name() != "groups" {
-			filtered = append(filtered, child)
-		}
-	}
-	t.Children = filtered
-
-	return nil
-}
-
-// FormatInheritance returns the config with inherited groups expanded and
-// annotated with "## 'X' was inherited from group 'Y'" comments, matching
-// Junos "show configuration | display inheritance" output.
-func (t *ConfigTree) FormatInheritance() string {
-	clone := t.Clone()
-	if err := clone.ExpandGroupsTagged(); err != nil {
-		return t.Format() // fallback to plain format on error
-	}
-	var b strings.Builder
-	formatNodesInheritance(&b, clone.Children, 0)
-	return b.String()
-}
-
 // navigatePath walks the tree following path components and returns matching nodes.
 // When multiple sibling nodes share the same key prefix (e.g., path ["from-zone","untrust"]
 // matching both ["from-zone","untrust","to-zone","trust"] and
@@ -292,277 +203,6 @@ func navigatePath(nodes []*Node, path []string) []*Node {
 		}
 	}
 	return nil
-}
-
-// FormatPathInheritance is like FormatPath but with inheritance annotations.
-func (t *ConfigTree) FormatPathInheritance(path []string) string {
-	clone := t.Clone()
-	if err := clone.ExpandGroupsTagged(); err != nil {
-		return t.FormatPath(path)
-	}
-	if len(path) == 0 {
-		return clone.FormatInheritance()
-	}
-	matches := navigatePath(clone.Children, path)
-	if len(matches) == 0 {
-		return ""
-	}
-	var b strings.Builder
-	for _, n := range matches {
-		if n.IsLeaf {
-			fmt.Fprintf(&b, "%s;\n", n.QuotedKeyPath())
-		} else {
-			fmt.Fprintf(&b, "%s {\n", n.QuotedKeyPath())
-			formatNodesInheritance(&b, n.Children, 1)
-			fmt.Fprintf(&b, "}\n")
-		}
-	}
-	return b.String()
-}
-
-func formatNodesInheritance(b *strings.Builder, nodes []*Node, indent int) {
-	nodes = canonicalOrder(nodes)
-	prefix := strings.Repeat("    ", indent)
-	for _, n := range nodes {
-		if n.Annotation != "" {
-			fmt.Fprintf(b, "%s/* %s */\n", prefix, n.Annotation)
-		}
-		if n.InheritedFrom != "" {
-			// Determine the display name for the annotation.
-			// Junos uses the last key for leaf values ("## 'any' was inherited")
-			// and the first non-keyword key for containers ("## 'default-deny' was inherited").
-			displayKey := n.Keys[len(n.Keys)-1]
-			fmt.Fprintf(b, "%s##\n%s## '%s' was inherited from group '%s'\n%s##\n",
-				prefix, prefix, displayKey, n.InheritedFrom, prefix)
-		}
-		if n.IsLeaf {
-			fmt.Fprintf(b, "%s%s;\n", prefix, n.QuotedKeyPath())
-		} else {
-			fmt.Fprintf(b, "%s%s {\n", prefix, n.QuotedKeyPath())
-			formatNodesInheritance(b, n.Children, indent+1)
-			fmt.Fprintf(b, "%s}\n", prefix)
-		}
-	}
-}
-
-// tagNodesInherited recursively sets InheritedFrom on all nodes.
-func tagNodesInherited(nodes []*Node, groupName string) {
-	for _, n := range nodes {
-		n.InheritedFrom = groupName
-		tagNodesInherited(n.Children, groupName)
-	}
-}
-
-// stripApplyGroups recursively removes apply-groups nodes and returns error
-// if they reference undefined groups. vars is used to resolve ${var}
-// placeholders in group names for error messages.
-func (t *ConfigTree) stripApplyGroups(vars map[string]string) error {
-	return stripApplyGroupsInNodes(t.Children, vars)
-}
-
-func stripApplyGroupsInNodes(nodes []*Node, vars map[string]string) error {
-	for _, child := range nodes {
-		if child.Name() == "apply-groups" {
-			name := ""
-			if len(child.Keys) > 1 {
-				name = resolveVars(child.Keys[1], vars)
-			}
-			return fmt.Errorf("apply-groups references undefined group %q", name)
-		}
-		if !child.IsLeaf {
-			if err := stripApplyGroupsInNodes(child.Children, vars); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// walkGroupToContext walks a group definition's tree to match the ancestor
-// context path. Each element of ancestorPath is the Keys slice of a parent
-// node from root to the current level. Returns the children of the deepest
-// matching node, or nil if the group has no matching subtree.
-// Supports <*> wildcard matching in group keys.
-func walkGroupToContext(groupChildren []*Node, ancestorPath [][]string) []*Node {
-	current := groupChildren
-	for _, pathKeys := range ancestorPath {
-		var next []*Node
-		for _, child := range current {
-			if child.IsLeaf {
-				continue
-			}
-			// Exact match or wildcard match (group keys may contain <*>).
-			if keysEqual(child.Keys, pathKeys) || keysMatchWildcard(pathKeys, child.Keys) {
-				next = child.Children
-				break
-			}
-		}
-		if next == nil {
-			return nil // group doesn't have matching subtree at this context
-		}
-		current = next
-	}
-	return current
-}
-
-// expandGroupsRecursive processes apply-groups nodes within a node list,
-// then recurses into all children to handle nested apply-groups.
-// ancestorPath tracks the key path from root to the current level, enabling
-// groups to be walked down to the matching context for nested apply-groups.
-// seen tracks group names being expanded to detect circular references.
-// If tagInherited is true, merged nodes get InheritedFrom set to the group name.
-// vars provides ${var} replacements for group names (may be nil).
-func expandGroupsRecursive(nodes *[]*Node, groups map[string]*Node, ancestorPath [][]string, seen map[string]bool, tagInherited bool, vars map[string]string) error {
-	// First, collect apply-groups references at this level.
-	// Support bracket-list syntax: apply-groups [ name1 name2 ] produces
-	// Keys = ["apply-groups", "name1", "name2"].
-	var applyNames []string
-	for _, n := range *nodes {
-		if n.Name() == "apply-groups" {
-			for _, key := range n.Keys[1:] {
-				applyNames = append(applyNames, resolveVars(key, vars))
-			}
-		}
-	}
-
-	// Expand each referenced group.
-	for _, name := range applyNames {
-		g, ok := groups[name]
-		if !ok {
-			return fmt.Errorf("apply-groups references undefined group %q", name)
-		}
-
-		if seen == nil {
-			seen = make(map[string]bool)
-		}
-		if seen[name] {
-			return fmt.Errorf("apply-groups circular reference: group %q", name)
-		}
-		seen[name] = true
-
-		// Walk the group tree to match the current context path.
-		var srcChildren []*Node
-		if len(ancestorPath) == 0 {
-			// Top-level: merge group's direct children.
-			srcChildren = g.Children
-		} else {
-			srcChildren = walkGroupToContext(g.Children, ancestorPath)
-		}
-
-		if srcChildren != nil {
-			cloned := cloneNodes(srcChildren)
-			if tagInherited {
-				tagNodesInherited(cloned, name)
-			}
-			mergeNodes(nodes, cloned)
-		}
-
-		delete(seen, name)
-	}
-
-	// Remove apply-groups nodes.
-	filtered := make([]*Node, 0, len(*nodes))
-	for _, n := range *nodes {
-		if n.Name() != "apply-groups" {
-			filtered = append(filtered, n)
-		}
-	}
-	*nodes = filtered
-
-	// Recurse into children to handle nested apply-groups.
-	for _, n := range *nodes {
-		if !n.IsLeaf && len(n.Children) > 0 {
-			childPath := make([][]string, len(ancestorPath)+1)
-			copy(childPath, ancestorPath)
-			childPath[len(ancestorPath)] = n.Keys
-			if err := expandGroupsRecursive(&n.Children, groups, childPath, seen, tagInherited, vars); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-// mergeNodes merges src nodes into dst. For container nodes with matching keys,
-// children are merged recursively. For leaf nodes or new containers, they are
-// appended (group values don't override existing explicit config — existing
-// config takes precedence via ordering, since the compiler uses first-match).
-func mergeNodes(dst *[]*Node, src []*Node) {
-	for _, s := range src {
-		if s.IsLeaf {
-			// Only add leaf if no matching leaf exists.
-			if !hasMatchingLeaf(*dst, s.Keys) {
-				*dst = append(*dst, s)
-			}
-			continue
-		}
-
-		// Check if source keys contain wildcards (<*>).
-		if keysContainWildcard(s.Keys) {
-			// Wildcard merge: apply to all matching containers in dst.
-			for _, d := range *dst {
-				if !d.IsLeaf && keysMatchWildcard(d.Keys, s.Keys) {
-					cloned := cloneNodes(s.Children)
-					mergeNodes(&d.Children, cloned)
-				}
-			}
-			continue
-		}
-
-		// Container node: find matching container in dst.
-		found := false
-		for _, d := range *dst {
-			if !d.IsLeaf && keysEqual(d.Keys, s.Keys) {
-				// Merge children recursively.
-				mergeNodes(&d.Children, s.Children)
-				found = true
-				break
-			}
-		}
-		if !found {
-			*dst = append(*dst, s)
-		}
-	}
-}
-
-// keysContainWildcard returns true if any key is the Junos wildcard "<*>".
-func keysContainWildcard(keys []string) bool {
-	for _, k := range keys {
-		if k == "<*>" {
-			return true
-		}
-	}
-	return false
-}
-
-// keysMatchWildcard checks if dst keys match src keys where "<*>" matches
-// any value. Both slices must have the same length.
-func keysMatchWildcard(dst, src []string) bool {
-	if len(dst) != len(src) {
-		return false
-	}
-	for i := range src {
-		if src[i] != "<*>" && src[i] != dst[i] {
-			return false
-		}
-	}
-	return true
-}
-
-// hasMatchingLeaf returns true if nodes contains a leaf whose first key
-// matches. This prevents group values from overriding explicit config
-// (e.g., if "host-name explicit" already exists, "host-name group" is skipped).
-func hasMatchingLeaf(nodes []*Node, keys []string) bool {
-	if len(keys) == 0 {
-		return false
-	}
-	for _, n := range nodes {
-		if n.IsLeaf && len(n.Keys) > 0 && n.Keys[0] == keys[0] {
-			return true
-		}
-	}
-	return false
 }
 
 // matchNodeKeys checks if a node's Keys match path elements starting at pos.
@@ -709,123 +349,6 @@ func (t *ConfigTree) findNodeWithParent(path []string) (*Node, *[]*Node, error) 
 		pos += bestConsumed
 	}
 	return nil, nil, fmt.Errorf("path not found")
-}
-
-// CopyPath copies a subtree from src to dst path.
-// The destination's last N keys (where N = len(sourceNode.Keys)) replace the source keys.
-func (t *ConfigTree) CopyPath(src, dst []string) error {
-	if len(src) == 0 || len(dst) == 0 {
-		return fmt.Errorf("empty path")
-	}
-	srcNode, _, err := t.findNodeWithParent(src)
-	if err != nil {
-		return fmt.Errorf("source not found: %s", strings.Join(src, " "))
-	}
-	cloned := cloneNodes([]*Node{srcNode})[0]
-	nk := len(srcNode.Keys)
-	if len(dst) < nk {
-		return fmt.Errorf("destination path too short")
-	}
-	cloned.Keys = append([]string(nil), dst[len(dst)-nk:]...)
-	// Find the parent for the destination
-	dstParentPath := dst[:len(dst)-nk]
-	return t.insertNode(dstParentPath, cloned)
-}
-
-// RenamePath moves a subtree from src to dst path.
-func (t *ConfigTree) RenamePath(src, dst []string) error {
-	if len(src) == 0 || len(dst) == 0 {
-		return fmt.Errorf("empty path")
-	}
-	srcNode, err := t.removeNode(src)
-	if err != nil {
-		return fmt.Errorf("source not found: %s", strings.Join(src, " "))
-	}
-	nk := len(srcNode.Keys)
-	if len(dst) < nk {
-		return fmt.Errorf("destination path too short")
-	}
-	srcNode.Keys = append([]string(nil), dst[len(dst)-nk:]...)
-	dstParentPath := dst[:len(dst)-nk]
-	return t.insertNode(dstParentPath, srcNode)
-}
-
-// InsertBefore moves an existing element before another element in the same
-// parent's children list. Both elements must exist under the same parent.
-// elementPath is the full path to the element to move.
-// refPath is the full path to the reference element.
-func (t *ConfigTree) InsertBefore(elementPath, refPath []string) error {
-	return t.insertRelative(elementPath, refPath, false)
-}
-
-// InsertAfter moves an existing element after another element in the same
-// parent's children list.
-func (t *ConfigTree) InsertAfter(elementPath, refPath []string) error {
-	return t.insertRelative(elementPath, refPath, true)
-}
-
-func (t *ConfigTree) insertRelative(elementPath, refPath []string, after bool) error {
-	if len(elementPath) == 0 || len(refPath) == 0 {
-		return fmt.Errorf("empty path")
-	}
-
-	// Find the element and its parent children slice.
-	elem, parentChildren, err := t.findNodeWithParent(elementPath)
-	if err != nil {
-		return fmt.Errorf("element not found: %s", strings.Join(elementPath, " "))
-	}
-
-	// Find the reference element — must be in the same parent.
-	refNode, refParent, err := t.findNodeWithParent(refPath)
-	if err != nil {
-		return fmt.Errorf("reference not found: %s", strings.Join(refPath, " "))
-	}
-
-	// Both must share the same parent (same children slice).
-	if parentChildren != refParent {
-		return fmt.Errorf("elements must be siblings (same parent)")
-	}
-
-	if elem == refNode {
-		return nil // already in position
-	}
-
-	// Find and remove the element from the children slice.
-	elemIdx := -1
-	for i, c := range *parentChildren {
-		if c == elem {
-			elemIdx = i
-			break
-		}
-	}
-	if elemIdx < 0 {
-		return fmt.Errorf("element not found in parent")
-	}
-	*parentChildren = append((*parentChildren)[:elemIdx], (*parentChildren)[elemIdx+1:]...)
-
-	// Find the reference's new index (may have shifted after removal).
-	refIdx := -1
-	for i, c := range *parentChildren {
-		if c == refNode {
-			refIdx = i
-			break
-		}
-	}
-	if refIdx < 0 {
-		return fmt.Errorf("reference not found in parent")
-	}
-
-	// Insert before or after the reference.
-	insertAt := refIdx
-	if after {
-		insertAt = refIdx + 1
-	}
-
-	// Grow and shift.
-	*parentChildren = append(*parentChildren, nil)
-	copy((*parentChildren)[insertAt+1:], (*parentChildren)[insertAt:])
-	(*parentChildren)[insertAt] = elem
-	return nil
 }
 
 // ValueHint identifies what kind of dynamic value is expected at a schema position.
@@ -2027,269 +1550,6 @@ func init() {
 	}
 }
 
-// SetPath inserts a leaf node at the given path in the tree.
-// Intermediate block nodes are created as needed. The schema determines
-// which keywords are containers (and how many extra args they consume)
-// versus leaves (all remaining tokens form the leaf's Keys).
-func (t *ConfigTree) SetPath(path []string) error {
-	if len(path) == 0 {
-		return fmt.Errorf("empty path")
-	}
-
-	current := &t.Children
-	schema := setSchema
-	i := 0
-
-	for i < len(path) {
-		keyword := path[i]
-
-		// Look up keyword in current schema level.
-		var childSchema *schemaNode
-		if schema != nil {
-			if s, ok := schema.children[keyword]; ok {
-				childSchema = s
-			} else if schema.wildcard != nil {
-				childSchema = schema.wildcard
-			}
-		}
-
-		if childSchema == nil {
-			// No schema match: all remaining tokens form a leaf node.
-			// Skip if exact duplicate already exists.
-			remaining := path[i:]
-			for _, n := range *current {
-				if n.IsLeaf && keysEqual(n.Keys, remaining) {
-					return nil
-				}
-			}
-			leaf := &Node{
-				Keys:   append([]string(nil), remaining...),
-				IsLeaf: true,
-			}
-			*current = append(*current, leaf)
-			return nil
-		}
-
-		// Consume keyword + extra args as this node's keys.
-		nodeKeyCount := 1 + childSchema.args
-		if i+nodeKeyCount > len(path) {
-			// Not enough tokens; treat remainder as leaf.
-			leaf := &Node{
-				Keys:   append([]string(nil), path[i:]...),
-				IsLeaf: true,
-			}
-			*current = append(*current, leaf)
-			return nil
-		}
-
-		nodeKeys := path[i : i+nodeKeyCount]
-		i += nodeKeyCount
-
-		// Compound key: children form part of the key rather than
-		// separate hierarchy levels (e.g. "family inet6" is a single
-		// node with Keys=["family","inet6"], not nested nodes).
-		if childSchema.compoundKey && i < len(path) {
-			if sub, ok := childSchema.children[path[i]]; ok {
-				nodeKeys = append(append([]string(nil), nodeKeys...), path[i])
-				i++
-				childSchema = sub
-			}
-		}
-
-		if i >= len(path) {
-			// No more tokens after this node: it's a leaf.
-			if childSchema.args > 0 && !childSchema.multi && childSchema.children == nil {
-				// Single-value leaf with no sub-structure (e.g. host-name, description): replace existing.
-				// Nodes with children are named containers that may appear as terminal leaves
-				// with different values (e.g. "interface eth0", "interface eth1").
-				// Replace the first match and remove all subsequent duplicates.
-				replaced := false
-				filtered := (*current)[:0] // reuse backing array
-				for _, n := range *current {
-					if n.IsLeaf && len(n.Keys) > 0 && n.Keys[0] == nodeKeys[0] {
-						if !replaced {
-							filtered = append(filtered, &Node{
-								Keys:   append([]string(nil), nodeKeys...),
-								IsLeaf: true,
-							})
-							replaced = true
-						}
-						// skip all duplicate entries
-						continue
-					}
-					filtered = append(filtered, n)
-				}
-				if replaced {
-					*current = filtered
-					return nil
-				}
-			} else {
-				// Flag leaf (args == 0) or multi-value leaf: skip if exact duplicate.
-				for _, n := range *current {
-					if n.IsLeaf && keysEqual(n.Keys, nodeKeys) {
-						return nil
-					}
-				}
-			}
-			leaf := &Node{
-				Keys:   append([]string(nil), nodeKeys...),
-				IsLeaf: true,
-			}
-			*current = append(*current, leaf)
-			return nil
-		}
-
-		// More tokens follow. If the schema says this is a multi-value leaf
-		// with no children AND the next token is a known sibling keyword,
-		// add it as a leaf and continue at the same level so remaining
-		// tokens become siblings (e.g. "match" children:
-		// destination-address any source-address any application any).
-		// If the next token is NOT a known sibling, the remaining tokens
-		// are trailing values for this leaf (e.g. "destination-port 20000 to 20003").
-		if childSchema.children == nil && childSchema.multi && i < len(path) {
-			nextToken := path[i]
-			_, nextIsSibling := schema.children[nextToken]
-			if !nextIsSibling && schema.wildcard != nil {
-				nextIsSibling = true
-			}
-			if nextIsSibling {
-				// Dedup: skip if exact leaf already exists.
-				dup := false
-				for _, n := range *current {
-					if n.IsLeaf && keysEqual(n.Keys, nodeKeys) {
-						dup = true
-						break
-					}
-				}
-				if !dup {
-					*current = append(*current, &Node{
-						Keys:   append([]string(nil), nodeKeys...),
-						IsLeaf: true,
-					})
-				}
-				// Don't descend — continue at same level for next sibling.
-				continue
-			}
-		}
-
-		// This is a container (or a leaf with trailing value tokens).
-		// Find or create matching node.
-		found := false
-		for _, n := range *current {
-			if !n.IsLeaf && keysEqual(n.Keys, nodeKeys) {
-				current = &n.Children
-				found = true
-				break
-			}
-		}
-		if !found {
-			newNode := &Node{
-				Keys: append([]string(nil), nodeKeys...),
-			}
-			*current = append(*current, newNode)
-			current = &newNode.Children
-		}
-		schema = childSchema
-	}
-
-	return nil
-}
-
-// DeletePath removes a node at the given path from the tree.
-// Uses the same schema-driven traversal as SetPath to navigate the tree,
-// then removes the target node from its parent's Children slice.
-func (t *ConfigTree) DeletePath(path []string) error {
-	if len(path) == 0 {
-		return fmt.Errorf("empty path")
-	}
-
-	return deletePath(&t.Children, path, setSchema, 0)
-}
-
-func deletePath(current *[]*Node, path []string, schema *schemaNode, i int) error {
-	if i >= len(path) {
-		return fmt.Errorf("path not found")
-	}
-
-	keyword := path[i]
-
-	// Look up keyword in current schema level.
-	var childSchema *schemaNode
-	if schema != nil {
-		if s, ok := schema.children[keyword]; ok {
-			childSchema = s
-		} else if schema.wildcard != nil {
-			childSchema = schema.wildcard
-		}
-	}
-
-	if childSchema == nil {
-		// No schema match: remaining tokens form leaf keys, remove matching node.
-		leafKeys := path[i:]
-		return removeMatchingNode(current, leafKeys)
-	}
-
-	// Consume keyword + extra args as this node's keys.
-	nodeKeyCount := 1 + childSchema.args
-	if i+nodeKeyCount > len(path) {
-		// Not enough tokens; treat remainder as leaf keys.
-		leafKeys := path[i:]
-		return removeMatchingNode(current, leafKeys)
-	}
-
-	nodeKeys := path[i : i+nodeKeyCount]
-	i += nodeKeyCount
-
-	// Compound key: consume child token as part of key.
-	if childSchema.compoundKey && i < len(path) {
-		if sub, ok := childSchema.children[path[i]]; ok {
-			nodeKeys = append(append([]string(nil), nodeKeys...), path[i])
-			i++
-			childSchema = sub
-		}
-	}
-
-	if i >= len(path) {
-		// No more tokens: this container node itself is the target.
-		return removeMatchingNode(current, nodeKeys)
-	}
-
-	// More tokens remain: find matching container and descend.
-	for _, n := range *current {
-		if !n.IsLeaf && keysEqual(n.Keys, nodeKeys) {
-			return deletePath(&n.Children, path, childSchema, i)
-		}
-	}
-
-	return fmt.Errorf("path not found: container %q does not exist", strings.Join(nodeKeys, " "))
-}
-
-// removeMatchingNode removes the first node whose keys match targetKeys
-// (using prefix matching) from the nodes slice.
-func removeMatchingNode(nodes *[]*Node, targetKeys []string) error {
-	for i, n := range *nodes {
-		if keysMatch(n.Keys, targetKeys) {
-			*nodes = append((*nodes)[:i], (*nodes)[i+1:]...)
-			return nil
-		}
-	}
-	return fmt.Errorf("path not found: no node matching %q", strings.Join(targetKeys, " "))
-}
-
-// keysMatch returns true if nodeKeys starts with all elements of targetKeys.
-// This allows "delete ... address srv1" to match a leaf ["address", "srv1", "10.0.1.0/32"].
-func keysMatch(nodeKeys, targetKeys []string) bool {
-	if len(targetKeys) > len(nodeKeys) {
-		return false
-	}
-	for i, tk := range targetKeys {
-		if nodeKeys[i] != tk {
-			return false
-		}
-	}
-	return true
-}
-
 // keysEqual returns true if two key slices are identical.
 func keysEqual(a, b []string) bool {
 	if len(a) != len(b) {
@@ -2455,460 +1715,4 @@ func CompleteSetPathWithValues(tokens []string, provider ValueProvider) []Schema
 		return nil
 	}
 	return completions
-}
-
-// Format renders the tree as Junos hierarchical configuration text.
-func (t *ConfigTree) Format() string {
-	var b strings.Builder
-	formatNodes(&b, t.Children, 0)
-	return b.String()
-}
-
-// canonicalOrder reorders children so "match"/"from" comes before "then",
-// matching Junos canonical display order for policies, NAT rules, and
-// firewall filter terms.
-func canonicalOrder(nodes []*Node) []*Node {
-	matchIdx, thenIdx := -1, -1
-	for i, n := range nodes {
-		if len(n.Keys) > 0 {
-			switch n.Keys[0] {
-			case "match", "from":
-				matchIdx = i
-			case "then":
-				thenIdx = i
-			}
-		}
-	}
-	if matchIdx < 0 || thenIdx < 0 || matchIdx < thenIdx {
-		return nodes // already correct or doesn't apply
-	}
-	// match/from is after then — move it to just before then
-	result := make([]*Node, 0, len(nodes))
-	for i, n := range nodes {
-		if i == matchIdx {
-			continue
-		}
-		if i == thenIdx {
-			result = append(result, nodes[matchIdx])
-		}
-		result = append(result, n)
-	}
-	return result
-}
-
-func formatNodes(b *strings.Builder, nodes []*Node, indent int) {
-	nodes = canonicalOrder(nodes)
-	prefix := strings.Repeat("    ", indent)
-	for _, n := range nodes {
-		if n.Annotation != "" {
-			fmt.Fprintf(b, "%s/* %s */\n", prefix, n.Annotation)
-		}
-		if n.IsLeaf {
-			fmt.Fprintf(b, "%s%s;\n", prefix, n.QuotedKeyPath())
-		} else {
-			fmt.Fprintf(b, "%s%s {\n", prefix, n.QuotedKeyPath())
-			formatNodes(b, n.Children, indent+1)
-			fmt.Fprintf(b, "%s}\n", prefix)
-		}
-	}
-}
-
-// FormatPath navigates to the given path and formats the subtree found there.
-// Path components are matched against node keys. For example, FormatPath(["interfaces", "wan0"])
-// navigates into the "interfaces" node, then into the child whose second key is "wan0",
-// and formats that subtree. Returns "" if the path is not found.
-func (t *ConfigTree) FormatPath(path []string) string {
-	if len(path) == 0 {
-		return t.Format()
-	}
-	matches := navigatePath(t.Children, path)
-	if len(matches) == 0 {
-		return ""
-	}
-	var b strings.Builder
-	for _, n := range matches {
-		if n.IsLeaf {
-			fmt.Fprintf(&b, "%s;\n", n.QuotedKeyPath())
-		} else {
-			fmt.Fprintf(&b, "%s {\n", n.QuotedKeyPath())
-			formatNodes(&b, n.Children, 1)
-			fmt.Fprintf(&b, "}\n")
-		}
-	}
-	return b.String()
-}
-
-// FormatSet renders the tree as flat "set" commands.
-func (t *ConfigTree) FormatSet() string {
-	var b strings.Builder
-	formatSetNodes(&b, t.Children, nil)
-	return b.String()
-}
-
-// FormatPathSet renders a subtree as flat "set" commands.
-// The full path prefix (including parent keys) is included in each set line.
-func (t *ConfigTree) FormatPathSet(path []string) string {
-	if len(path) == 0 {
-		return t.FormatSet()
-	}
-	matches := navigatePath(t.Children, path)
-	if len(matches) == 0 {
-		return ""
-	}
-	// Compute parent prefix: path elements before the matched node's first key.
-	var parentPrefix []string
-	firstKey := matches[0].Keys[0]
-	for _, p := range path {
-		if p == firstKey {
-			break
-		}
-		parentPrefix = append(parentPrefix, p)
-	}
-	var b strings.Builder
-	for _, n := range matches {
-		prefix := append(append([]string{}, parentPrefix...), n.Keys...)
-		if n.IsLeaf {
-			fmt.Fprintf(&b, "set %s\n", joinQuotedKeys(prefix))
-		} else {
-			formatSetNodes(&b, n.Children, prefix)
-		}
-	}
-	return b.String()
-}
-
-func formatSetNodes(b *strings.Builder, nodes []*Node, prefix []string) {
-	for _, n := range canonicalOrder(nodes) {
-		path := append(prefix, n.Keys...)
-		if n.IsLeaf {
-			fmt.Fprintf(b, "set %s\n", joinQuotedKeys(path))
-		} else {
-			formatSetNodes(b, n.Children, path)
-		}
-	}
-}
-
-// joinQuotedKeys joins keys with spaces, quoting any that contain special characters.
-func joinQuotedKeys(keys []string) string {
-	parts := make([]string, len(keys))
-	for i, k := range keys {
-		parts[i] = quoteKey(k)
-	}
-	return strings.Join(parts, " ")
-}
-
-// FormatCompare produces a Junos-style hierarchical diff between two trees.
-// It shows [edit <path>] context headers and +/- prefixed lines for added/removed content.
-// Unchanged sibling nodes within changed containers are shown collapsed as "name { ... }".
-func FormatCompare(oldTree, newTree *ConfigTree) string {
-	var b strings.Builder
-	diffNodes(&b, oldTree.Children, newTree.Children, nil)
-	return b.String()
-}
-
-// diffNodes compares two sets of children at the same tree level.
-// It recurses into modified containers to find the deepest [edit] context,
-// only showing siblings at the level where actual leaf changes occur.
-func diffNodes(b *strings.Builder, oldNodes, newNodes []*Node, editPath []string) {
-	oldByKey := make(map[string]*Node, len(oldNodes))
-	for _, n := range oldNodes {
-		oldByKey[n.KeyPath()] = n
-	}
-	newByKey := make(map[string]*Node, len(newNodes))
-	for _, n := range newNodes {
-		newByKey[n.KeyPath()] = n
-	}
-
-	// Collect changed entries at this level.
-	type diffEntry struct {
-		oldNode *Node
-		newNode *Node
-	}
-
-	// Use canonical ordering from new tree, then appended removed-only entries.
-	seen := make(map[string]bool)
-	var entries []diffEntry
-	for _, n := range canonicalOrder(newNodes) {
-		kp := n.KeyPath()
-		seen[kp] = true
-		old := oldByKey[kp]
-		if old == nil || !nodesEqual(old, n) {
-			entries = append(entries, diffEntry{oldNode: old, newNode: n})
-		}
-	}
-	for _, n := range canonicalOrder(oldNodes) {
-		kp := n.KeyPath()
-		if !seen[kp] {
-			entries = append(entries, diffEntry{oldNode: n, newNode: nil})
-		}
-	}
-
-	if len(entries) == 0 {
-		return
-	}
-
-	// Check if all changes can be recursed into (both old and new are blocks).
-	// If so, recurse without printing siblings at this level.
-	allRecursable := true
-	for _, e := range entries {
-		if e.oldNode == nil || e.newNode == nil {
-			allRecursable = false
-			break
-		}
-		if e.oldNode.IsLeaf || e.newNode.IsLeaf {
-			allRecursable = false
-			break
-		}
-	}
-
-	if allRecursable {
-		// All changes are in modified sub-containers — recurse deeper without showing this level.
-		for _, e := range entries {
-			childPath := append(append([]string{}, editPath...), strings.Fields(e.oldNode.KeyPath())...)
-			diffNodes(b, e.oldNode.Children, e.newNode.Children, childPath)
-		}
-		return
-	}
-
-	// Print [edit <path>] header — this is the level where leaf changes exist.
-	if len(editPath) > 0 {
-		fmt.Fprintf(b, "[edit %s]\n", strings.Join(editPath, " "))
-	}
-
-	// Show all children at this level: unchanged as collapsed, added/removed with prefix.
-	// Merge all nodes from both old and new in canonical order (new first, then old-only).
-	seen2 := make(map[string]bool)
-	var allEntries []diffEntry
-	for _, n := range canonicalOrder(newNodes) {
-		kp := n.KeyPath()
-		seen2[kp] = true
-		allEntries = append(allEntries, diffEntry{oldNode: oldByKey[kp], newNode: n})
-	}
-	for _, n := range canonicalOrder(oldNodes) {
-		if !seen2[n.KeyPath()] {
-			allEntries = append(allEntries, diffEntry{oldNode: n, newNode: nil})
-		}
-	}
-
-	indent := "    "
-	for _, e := range allEntries {
-		switch {
-		case e.oldNode == nil:
-			// Added
-			formatPrefixed(b, "+", indent, e.newNode)
-		case e.newNode == nil:
-			// Removed
-			formatPrefixed(b, "-", indent, e.oldNode)
-		case nodesEqual(e.oldNode, e.newNode):
-			// Unchanged — show collapsed
-			if e.oldNode.IsLeaf {
-				fmt.Fprintf(b, " %s%s;\n", indent, e.oldNode.QuotedKeyPath())
-			} else {
-				fmt.Fprintf(b, " %s%s { ... }\n", indent, e.oldNode.QuotedKeyPath())
-			}
-		default:
-			// Modified
-			if !e.oldNode.IsLeaf && !e.newNode.IsLeaf {
-				// Both are blocks — show [edit] context for sub-container
-				childPath := append(append([]string{}, editPath...), strings.Fields(e.oldNode.KeyPath())...)
-				diffNodes(b, e.oldNode.Children, e.newNode.Children, childPath)
-			} else {
-				formatPrefixed(b, "-", indent, e.oldNode)
-				formatPrefixed(b, "+", indent, e.newNode)
-			}
-		}
-	}
-}
-
-// nodesEqual returns true if two nodes have identical content (deep comparison).
-func nodesEqual(a, b *Node) bool {
-	if a.KeyPath() != b.KeyPath() {
-		return false
-	}
-	if a.IsLeaf != b.IsLeaf {
-		return false
-	}
-	if a.IsLeaf {
-		return true
-	}
-	if len(a.Children) != len(b.Children) {
-		return false
-	}
-	bByKey := make(map[string]*Node, len(b.Children))
-	for _, n := range b.Children {
-		bByKey[n.KeyPath()] = n
-	}
-	for _, ac := range a.Children {
-		bc, ok := bByKey[ac.KeyPath()]
-		if !ok {
-			return false
-		}
-		if !nodesEqual(ac, bc) {
-			return false
-		}
-	}
-	return true
-}
-
-// formatPrefixed writes a node with +/- prefix at the given indent.
-func formatPrefixed(b *strings.Builder, prefix, indent string, n *Node) {
-	if n.IsLeaf {
-		fmt.Fprintf(b, "%s%s%s;\n", prefix, indent, n.QuotedKeyPath())
-	} else {
-		fmt.Fprintf(b, "%s%s%s {\n", prefix, indent, n.QuotedKeyPath())
-		formatPrefixedChildren(b, prefix, indent+"    ", n.Children)
-		fmt.Fprintf(b, "%s%s}\n", prefix, indent)
-	}
-}
-
-// formatPrefixedChildren writes all children with the same +/- prefix.
-func formatPrefixedChildren(b *strings.Builder, prefix, indent string, nodes []*Node) {
-	for _, n := range canonicalOrder(nodes) {
-		formatPrefixed(b, prefix, indent, n)
-	}
-}
-
-// FormatJSON renders the tree as a JSON object.
-func (t *ConfigTree) FormatJSON() string {
-	obj := nodesToJSON(t.Children)
-	data, err := json.MarshalIndent(obj, "", "  ")
-	if err != nil {
-		return "{}"
-	}
-	return string(data) + "\n"
-}
-
-// FormatPathJSON renders a subtree as a JSON object.
-func (t *ConfigTree) FormatPathJSON(path []string) string {
-	if len(path) == 0 {
-		return t.FormatJSON()
-	}
-	matches := navigatePath(t.Children, path)
-	if len(matches) == 0 {
-		return ""
-	}
-	obj := nodesToJSON(matches)
-	data, err := json.MarshalIndent(obj, "", "  ")
-	if err != nil {
-		return "{}"
-	}
-	return string(data) + "\n"
-}
-
-// FormatXML renders the tree as Junos-style XML configuration.
-func (t *ConfigTree) FormatXML() string {
-	var b strings.Builder
-	b.WriteString(xml.Header)
-	b.WriteString("<configuration>\n")
-	formatXMLNodes(&b, t.Children, 1)
-	b.WriteString("</configuration>\n")
-	return b.String()
-}
-
-// FormatPathXML renders a subtree as Junos-style XML.
-func (t *ConfigTree) FormatPathXML(path []string) string {
-	if len(path) == 0 {
-		return t.FormatXML()
-	}
-	matches := navigatePath(t.Children, path)
-	if len(matches) == 0 {
-		return ""
-	}
-	var b strings.Builder
-	b.WriteString(xml.Header)
-	b.WriteString("<configuration>\n")
-	formatXMLNodes(&b, matches, 1)
-	b.WriteString("</configuration>\n")
-	return b.String()
-}
-
-func formatXMLNodes(b *strings.Builder, nodes []*Node, indent int) {
-	prefix := strings.Repeat("    ", indent)
-	for _, n := range nodes {
-		if n.IsLeaf {
-			formatXMLLeaf(b, n, prefix)
-		} else {
-			tag := xmlTag(n.Keys[0])
-			fmt.Fprintf(b, "%s<%s>\n", prefix, tag)
-			// Extra keys become <name> elements.
-			for _, k := range n.Keys[1:] {
-				fmt.Fprintf(b, "%s    <name>%s</name>\n", prefix, xmlEscape(k))
-			}
-			formatXMLNodes(b, n.Children, indent+1)
-			fmt.Fprintf(b, "%s</%s>\n", prefix, tag)
-		}
-	}
-}
-
-func formatXMLLeaf(b *strings.Builder, n *Node, prefix string) {
-	if len(n.Keys) == 1 {
-		// Boolean leaf: <keyword/>
-		fmt.Fprintf(b, "%s<%s/>\n", prefix, xmlTag(n.Keys[0]))
-		return
-	}
-	// Leaf with value: <keyword>value</keyword>
-	// For multi-key leaves like "address 10.0.1.0/24", emit
-	// <keyword><name>val1</name></keyword>
-	tag := xmlTag(n.Keys[0])
-	if len(n.Keys) == 2 {
-		fmt.Fprintf(b, "%s<%s>%s</%s>\n", prefix, tag, xmlEscape(n.Keys[1]), tag)
-	} else {
-		fmt.Fprintf(b, "%s<%s>\n", prefix, tag)
-		for _, k := range n.Keys[1:] {
-			fmt.Fprintf(b, "%s    <name>%s</name>\n", prefix, xmlEscape(k))
-		}
-		fmt.Fprintf(b, "%s</%s>\n", prefix, tag)
-	}
-}
-
-// xmlTag sanitizes a Junos keyword into a valid XML element name.
-func xmlTag(s string) string {
-	// Junos keywords already use valid XML chars (letters, digits, hyphens).
-	return s
-}
-
-// xmlEscape escapes special XML characters in text content.
-func xmlEscape(s string) string {
-	var b strings.Builder
-	xml.EscapeText(&b, []byte(s))
-	return b.String()
-}
-
-// nodesToJSON converts a list of AST nodes to a nested map structure.
-func nodesToJSON(nodes []*Node) map[string]interface{} {
-	result := make(map[string]interface{})
-
-	for _, n := range nodes {
-		if n.IsLeaf {
-			// Leaf node: key is first key, value is remaining keys joined
-			if len(n.Keys) == 1 {
-				result[n.Keys[0]] = true
-			} else if len(n.Keys) == 2 {
-				result[n.Keys[0]] = n.Keys[1]
-			} else {
-				result[n.Keys[0]] = strings.Join(n.Keys[1:], " ")
-			}
-		} else {
-			name := n.Keys[0]
-			qualifier := ""
-			if len(n.Keys) > 1 {
-				qualifier = strings.Join(n.Keys[1:], " ")
-			}
-
-			children := nodesToJSON(n.Children)
-
-			if qualifier != "" {
-				// Named instance: e.g. "interface trust0" → {"interface": {"trust0": {...}}}
-				if existing, ok := result[name]; ok {
-					if m, ok := existing.(map[string]interface{}); ok {
-						m[qualifier] = children
-					}
-				} else {
-					result[name] = map[string]interface{}{qualifier: children}
-				}
-			} else {
-				result[name] = children
-			}
-		}
-	}
-	return result
 }
